+++ CHANGELOG.md
@@ -4,13 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.6.0] - 2018-01-23
 ### Added
 - Mock-up test file.
 - Helpers test file.
 
 ### Changed
 - Router tests.
+- Structure stored in the radix tree (for performance reasons).
+- Key used to retrieve params in the request's context.
 
 ## [0.5.1] - 2017-12-21
 ### Fixed
@@ -76,7 +78,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - This package's source code, including examples and tests.
 - Go dep files.
 
-[Unreleased]: https://github.com/gbrlsnchs/httpmux/compare/v0.5.1...HEAD
+[0.6.0]: https://github.com/gbrlsnchs/httpmux/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/gbrlsnchs/httpmux/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/gbrlsnchs/httpmux/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/gbrlsnchs/httpmux/compare/v0.3.0...v0.4.0